First commit of basic recipie/ingredients parser
This first commit sets up the basic structure for the project with a recipies
module and associated structs to hold recipie data. Initially I've focused on
creating a parser for ingredients, using a sample recipie from the web.

The goal is to create a parser that will parse an entire recipie from Markdown
based text file and use that to define a generalized recipie specification.