Add group handling and use try_from for String -> Recipe
This commit adds the ability to store ingredients and instruction lists by
groups as well as switching to use try_from for converting Strings to Recipes.
This is to reflect the fact that if we don't get a good input string, we may not
be able to create a recipe.